# Speed up `Volume.electrondensity_spin` and `Volume.wavefunction`