Being able to remove explicit indexing and passing single points for calculating basis function amplitudes on a grid brought this code from 124 to 7 seconds, now almost completely in the Bader code:
```python
from pathlib import Path
from cclib.method import Bader, volume
from cclib.io import ccread
import numpy as np

path_data = Path("/home/eric/development/cclib_berquist/test/method/hf.out")
data = ccread(str(path_data))
vol = volume.Volume((-6, -6, -6), (6, 6, 6), (0.2, 0.2, 0.2))
analysis = Bader(data, vol)
analysis.calculate()
np.testing.assert_allclose(analysis.fragcharges, [10.58407202, 0.35639819], atol=1.0e-8)
```
and the non-regression test suite time from 547 to 214 seconds.

The current breakdown for non-regression test times is:
```
55.73s call     test_method.py::BaderTest::test_symms_benzene                                                                                                                
33.80s call     test_method.py::HirshfeldTest::test_water_charges                                                                                                            
18.47s call     test_method.py::HirshfeldTest::test_chgsum_co
7.84s call     test_method.py::DDEC6Test::test_water_charges
7.72s call     test_data.py::test_all[parsers0-modules0-True-False-40-True-True]
6.30s call     test_method.py::HirshfeldTest::test_chgsum_h2
5.72s call     test_method.py::BaderTest::test_chgsum_hf
5.33s call     test_method.py::DDEC6Test::test_chgsum_co
2.77s call     test_method.py::DDEC6Test::test_chg_nh3
1.68s call     test_method.py::BaderTest::test_val
```
and everything else took less than a second each.