@berquist 2020-10-05
I do want this for 1.6.4. I think it's fine as-is, but was trying to speed up the Python 2 implementation. Right now, Travis takes < 12 minutes for Python 3 and 15 minutes for Python 2. I think this is because the PyQuante 1 code isn't vectorized. I monkey-patched it with the vectorized implementation taken from PyQuante 2, only to find that the DDEC6 test for H2 never terminates: the loop inside https://github.com/cclib/cclib/blob/67be86feb8d97eb123c268b86323afec008d6dc5/cclib/method/ddec.py#L699 when called from `condition_densities` because it stays stuck at zero, where in the Python 3 implementation it's just barely above (~1e-17). I've managed to trace this back to the CGBFs themselves. There are two issues:
1. PyQuante 1 uses a different Angstrom-to-bohr conversion factor than in `cclib.utils.convertor`. I fixed this by removing the need for `cclib2pyquante` like this PR already did for PyQuante 2/Python 3.
2. The contraction coefficients are slightly different. Since they're used directly from `ccdata.gbasis`, the only explanation is that the routine which normalizes the contraction coefficients is different. There's actually a comment in the PyQuante 2 code wondering if it's correct. Though, this doesn't explain why switching to a vectorized method for putting basis functions on a grid would change the result to begin with.

Unless we want to speed up CI time for a language that will eventually be removed anyway, I won't debug this further and will just leave the implementation as-is.