@langner 2020-10-05
> Yeah it seems like wasted effort to try to optimize the python 2 further. I can review this later today.

Agreed. At this point I'm OK with _slowing down_ Python 2 :)